Tweak changelog
Technically we've not released a Debian package version yet, so we're
still UNRELEASED in -1, and I'm fixing up the changelog manually for
Ubuntu uploads. Doing it this way will make "dch" work correctly and my
Ubuntu uploads easier.

But in the long run we should either upload to Debian or decide we won't
and start tracking Ubuntu versions in master instead. This is probably a
discussion we should have on the pkg-mysql-maint list soon.